requirements.txt
scapy
python-whois